LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/unsupported-axis-expected.txt

Test

PASS (width > 0px)
PASS (height > 0px)
FAIL ((height > 0px) or (width > 0px)) assert_equals: expected "" but got "true"
FAIL ((width > 0px) or (height > 0px)) assert_equals: expected "" but got "true"
FAIL ((orientation: landscape) or (width > 0px)) assert_equals: expected "" but got "true"
FAIL ((width > 0px) or (orientation: landscape)) assert_equals: expected "" but got "true"
PASS ((height > 0px) or (orientation: landscape))
PASS ((height > 0px) or (orientation: landscape)), with contain:size
PASS (inline-size > 0px)
PASS (block-size > 0px)
PASS (block-size > 0px), with writing-mode:vertical-rl
